refactor: react의 stt 관련 utils 정리
